* Move handling of sig_restart_requested from mainwindow to Applicatioin.plugin * Preferences plugin handles user confirmation on sig_reset_preferences_requested * Preferences plugin handles SPYDER_RESET and emits sig_restart_requested
